new functions to generate particles
