+++ fma_ions/longitudinal/__init__.py
@@ -1,3 +1,4 @@
 from .parabolic_distribution import generate_parabolic_distribution
 from .binomial_distribution import generate_binomial_distribution
 from .binomial_distribution import generate_binomial_distribution_from_PS_extr
+from .longitudinal_particle_generator import generate_particles_transverse_gaussian, build_particles_linear_in_zeta, return_separatrix_coordinates
